requirements.txt
#pip install -r requirements.txt

# Modular RAG System Dependencies
# =================================

# PDF Processing (chunking_mejorado.py)
PyMuPDF==1.23.8

# Vector Embeddings & Search  
sentence-transformers>=2.2.0,<3.0.0
transformers>=4.21.0,<5.0.0
torch>=2.0.0
numpy>=1.24.0

# Vector Database (Qdrant)
qdrant-client>=1.6.0,<2.0.0

# Utilities
huggingface-hub>=0.16.0,<1.0.0

# ===== MODULE ARCHITECTURE =====
# rag_data_loader.py    → ETL: Load data to Qdrant
# search_engine.py      → Search: Semantic queries  
# test_searches.py      → Testing: Experiments & demos
# chunking_mejorado.py  → Processing: PDF → JSON

# Additional dependencies for stability
torch>=1.9.0
numpy>=1.21.0
requests>=2.25.0

# LLM Integration (OpenRouter)
openai>=1.0.0
python-dotenv>=1.0.0